Final commit ready for grading
I finished the project. I had to go back and fix the radio buttons because they weren't working and I also had to add in the buttons and the check box. I had a problem with the buttons receiving focus and I had to ask Conner for help and I was able to fix it